1.Fix the lawlFilter and average length method (from "int" to "double") 2.Fix a bug in getASCIIWord method which would go into infinite loop encountering some specific pucntuation.

+++ HandySearch/WordSegmenter.h
@@ -5,7 +5,7 @@ private:
 	QStringList words;
 	int wordCount;
 	int totalLen;
-	int avgLen;
+	double avgLen;
 	double variance;
 public:
 	Chunk();
@@ -13,7 +13,7 @@ public:
 	Chunk(QString fWord, QString sWord, QString tWord);
 	int getCount();
 	int getLength();
-	int getAvgLen();
+	double getAvgLen();
 	double getVariance();
 	QStringList &getWords();
 	Chunk &operator=(const Chunk &other);
